refactor : package 명view -> domain 으로 변경
