+++ src/main/java/univ/earthbreaker/namu/core/domain/character/book/MemberCharacter.java
@@ -1,4 +1,4 @@
-package univ.earthbreaker.namu.core.view.character;
+package univ.earthbreaker.namu.core.domain.character.book;
 
 import univ.earthbreaker.namu.core.domain.character.CharacterType;
 import univ.earthbreaker.namu.core.domain.character.Gender;
@@ -9,8 +9,8 @@ public class MemberCharacter {
 	private final long no;
 	private final long memberNo;
 	private final int count;
-	private final NamuCharacter character;
 	private final boolean isAcquired;
+	private final NamuCharacter character;
 
 	public MemberCharacter(long no, long memberNo, int count, NamuCharacter character, boolean isAcquired) {
 		this.no = no;
@@ -20,8 +20,24 @@ public class MemberCharacter {
 		this.isAcquired = isAcquired;
 	}
 
+	public static MemberCharacter notAcquired(long memberNo, NamuCharacter character) {
+		return new MemberCharacter(0, memberNo, 0, character, false);
+	}
+
+	public boolean isSameWithCharacterType(CharacterType type) {
+		return character.getType().equals(type);
+	}
+
 	public int getCount() {
 		return count;
+	}
+
+	public boolean isAcquired() {
+		return isAcquired;
+	}
+
+	public NamuCharacter getCharacter() {
+		return character;
 	}
 
 	public long getCharacterNo() {
@@ -50,5 +66,9 @@ public class MemberCharacter {
 
 	public String getDetailImagePath() {
 		return character.getDetailImagePath();
+	}
+
+	public String getThumbnailImagePath() {
+		return character.getMainImagePath();
 	}
 }
